azure-backend/requirements.txt
azure-functions==1.18.0
PyMuPDF==1.23.8
python-docx==1.1.0
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1
numpy==1.26.2
pandas==2.1.4
azure-cosmos==4.5.1
azure-storage-blob==12.19.0






